refactor: remove subpackages in auth package
