+++ backend/src/main/java/ch/sectioninformatique/template/auth/PasswordConfig.java
@@ -1,4 +1,4 @@
-package ch.sectioninformatique.template.auth.password;
+package ch.sectioninformatique.template.auth;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
