+++ backend/src/test/java/ch/sectioninformatique/template/user/UserServiceTest.java
@@ -1,7 +1,7 @@
 package ch.sectioninformatique.template.user;
 
-import ch.sectioninformatique.template.auth.credentials.CredentialsDto;
-import ch.sectioninformatique.template.auth.signup.SignUpDto;
+import ch.sectioninformatique.template.auth.CredentialsDto;
+import ch.sectioninformatique.template.auth.SignUpDto;
 import ch.sectioninformatique.template.app.exceptions.AppException;
 import ch.sectioninformatique.template.security.Role;
 import ch.sectioninformatique.template.security.RoleEnum;
